Add the timezone to Last Updated
Since currently, the webpage is fully statically rendered, the time may not properly reflect reality. The build server seems to be far ahead, which makes this timestamp appear in the future for some users (including me). The timezone will be listed beside the timestamp to put some context and perspective on when the page was actually last updated.