editor: minor refactor to only show a single item for performance
Before, this page showed every single item before hand, and this was a huge performance issue. When opening the page, nothing would render for a while, and I never bothered putting a loading spinner. Even if I did, editing anything felt sluggish, and really bad to use. It also was a pain to navigate, having to use Ctrl+F to find the items you were looking for. Now the page has a dropdown, and shows a single item at a time. Not only does this greatly increase performance, it's a lot easier to navigate and use. In the future, I might use the autocomplete dropdown to make this even easier to navigate!