editor: more concisely compare game object fields
When I was just using two different array values, this would not be very accurate if the list was heavily modified, or had a lot of deletes or insertions. To solve this, I map the Game arrays into Maps with their slugs as the key as those have to be unique, then get and compare with those when iterating. If there is a discrepancy, I can be assured that it's accurate and not a false positive.