games: added SYNCED
For why I didn't do this in the editor? Right now it doesn't support adding an anti-cheat that isn't on the list. How do you add an anti-cheat to the list? It already has to be there! It's a bit of a chicken and egg problem, and I think I could fix this by allow the input to accept any string, not just already pre-validated anti-cheats. But right now I'm in the middle of a refactor moving everything to Mantine v7, so it has to be done later.

Closes https://github.com/AreWeAntiCheatYet/AreWeAntiCheatYet/issues/1391.